stock-sentiment-system/requirements.txt
kafka-python
pyspark
transformers
torch
requests
yfinance
pandas
plotly
dash
python-dotenv
tweepy
